+++ src/parseFilename.js
@@ -1,0 +1,15 @@
+export default function parseFilename(filename) {
+    if (~filename.indexOf('.')) {
+        const [extensionWithoutDot, ...nameSegments] = filename.split('.').reverse();
+
+        return {
+            extname: '.' + extensionWithoutDot,
+            name: nameSegments.reverse().join('.')
+        };
+    } else {
+        return {
+            extname: '',
+            name: filename
+        };
+    }
+}