+++ CHANGELOG.md
@@ -64,6 +64,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Fix [#235](https://github.com/Microsoft/BotFramework-DirectLineJS/issues/235). Added metadata when uploading attachments, including `thumbnailUrl`, by [@compulim](https://github.com/compulim), in PR [#236](https://github.com/Microsoft/BotFramework-DirectLineJS/pull/236)
 
+### Fixed
+- Avoid posting an error on intentional end, by [@orgads](https://github.com/orgads) in PR [#172](https://github.com/Microsoft/BotFramework-DirectLineJS/pull/172)
+- Surface Web Socket errors, by [@orgads](https://github.com/orgads) in PR [#171](https://github.com/Microsoft/BotFramework-DirectLineJS/pull/171)
+
 ## [0.11.4] - 2019-03-04
 ### Changed
 - Change reconnect delay to be a random amount between 3s and 15s, by [@mingweiw](https://github.com/mingweiw) in PR [#164](https://github.com/Microsoft/BotFramework-DirectLineJS/pull/164)